Attempt switch to float64 for more 'precise' rounding

+++ modules/local/scanpy/bbknn/templates/bbknn.py
@@ -39,7 +39,7 @@ adata = bbknn.bbknn(adata, **kwargs)
 if params.decimals is not None:
     for key in adata.obsp:
         if hasattr(adata.obsp[key], 'data'):
-            adata.obsp[key].data = np.round(adata.obsp[key].data, params.decimals)
+            adata.obsp[key].data = np.round(adata.obsp[key].data.astype(np.float64), params.decimals)
 
 adata.write_h5ad("${prefix}.h5ad")
